+++ src/main/java/com/example/startapp/entity/ForgotPassword.java
@@ -15,7 +15,7 @@ public class ForgotPassword {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer fpid;
+    private Long fpid;
 
     @Column(nullable = false)
     private Integer otp;
